requirements.txt
scikit-learn
python-dotenv
feedparser
requests
openai
slack_sdk